Change: inlineSvelteComponent -> inlineComponent. Switch to ejswrapper
